Updated with animation
Created an animating loop, where each cycle is a different mod from mod
1 to mod 255. It also does not clear the last frame, allowing the
formula modulo 255 to be shown.

Added a sleep function using timeout, although there are definitely ways
to do this without having an async function.

Added a clear function, which clears the canvas and then draws a white
rectangle for background.